move mobile channel files to its own folder under mobile
